changed variable name, conditional render
the variable accAge is changed to more suitable name: joined.
GamesVoted - removed capital letter at start.
added conditional rendering for profile stats, so if one stat is not added it is not shown
for now the sample stat is hardcoded, this will be removed at later stages when certain elements are implemented